app/requirements.txt
# Core FastAPI and ASGI
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20

# Streamlit
streamlit>=1.48.0

# Data Processing
pandas>=2.3.0
numpy>=2.2.0
scikit-learn>=1.7.0

# Visualization
plotly>=6.3.0
matplotlib>=3.10.0
seaborn>=0.13.0

# Machine Learning
catboost>=1.2.0
torch>=2.8.0
lightgbm>=4.6.0
shap>=0.48.0

# Utilities
python-dotenv>=1.1.0
websockets>=15.0.0
httpx>=0.28.0
pydantic>=2.11.0

# Development
pytest>=8.4.0
black>=25.1.0
